signout action

+++ dashboard/src/actions/auth.action.js
@@ -2,8 +2,8 @@ import axios from "../helpers/axios"
 import { authConstants } from "./constants"
 
 export const login = (user) => {
-
     return async (dispatch) => {
+
         dispatch({ type: authConstants.LOGIN_REQUEST })
         const res = await axios.post('/admin/signin', {
             ...user
@@ -31,6 +31,8 @@ export const login = (user) => {
         }
     }
 }
+
+
 //keep token persistent in state 
 export const isUserLoggedIn = () => {
     return async dispatch => {
@@ -56,9 +58,20 @@ export const isUserLoggedIn = () => {
 
 export const signout = () => {
     return async (dispatch) => {
-        localStorage.clear();
-        dispatch({
-            type: authConstants.LOGOUT_REQUEST
-        })
+
+        dispatch({ type: authConstants.LOGIN_REQUEST })
+        const res = await axios.post('/admin/signout');
+        if (res.status === 200) {
+            localStorage.clear();
+            dispatch({ type: authConstants.LOGOUT_SUCCESS })
+        } else {
+            dispatch({
+                type: authConstants.LOGIN_FAILURE,
+                payload: {
+                    error: res.data.error
+                }
+            })
+        }
+
     }
 }
